fix: API circular imports; update code to address changes
