+++ packages/zowe-explorer-api/src/index.ts
@@ -9,15 +9,12 @@
  *
  */
 
+export * from "./extend";
 export * from "./profiles/UserSettings";
 export * from "./profiles/ProfilesCache";
-export * from "./profiles/ZoweExplorerApi";
 export * from "./profiles/ZoweExplorerZosmfApi";
 export * from "./logger/IZoweLogger";
-export * from "./tree/ZoweExplorerTreeApi";
-export * from "./tree/ZoweTreeNode";
-export * from "./tree/IZoweTree";
-export * from "./tree/IZoweTreeNode";
+export * from "./tree";
 export * from "./vscode/ZoweVsCodeExtension";
 export * from "./security/KeytarApi";
 export * from "./security/KeytarCredentialManager";
