+++ packages/zowe-explorer-api/src/tree/index.ts
@@ -9,7 +9,16 @@
  *
  */
 
+import { IZoweUSSTreeNode } from "./IZoweUSSTreeNode";
+import { IZoweDatasetTreeNode } from "./IZoweDatasetTreeNode";
+import { IZoweJobTreeNode } from "./IZoweJobTreeNode";
+
+export type IZoweNodeType = IZoweDatasetTreeNode | IZoweUSSTreeNode | IZoweJobTreeNode;
+
 export * from "./ZoweExplorerTreeApi";
 export * from "./ZoweTreeNode";
 export * from "./IZoweTree";
 export * from "./IZoweTreeNode";
+export * from "./IZoweDatasetTreeNode";
+export * from "./IZoweJobTreeNode";
+export * from "./IZoweUSSTreeNode";
