+++ packages/zowe-explorer/src/ZoweExplorerExtender.ts
@@ -16,17 +16,17 @@ import * as fs from "fs";
 import * as globals from "./globals";
 import * as vscode from "vscode";
 import {
+    IApiExplorerExtender,
+    getFullPath,
+    getZoweDir,
     Gui,
-    ZoweExplorerApi,
-    ZoweExplorerTreeApi,
     IZoweTree,
     IZoweTreeNode,
     IZoweDatasetTreeNode,
     IZoweUSSTreeNode,
     IZoweJobTreeNode,
     ProfilesCache,
-    getZoweDir,
-    getFullPath,
+    ZoweExplorerTreeApi,
 } from "@zowe/zowe-explorer-api";
 import { Profiles } from "./Profiles";
 import { ZoweExplorerApiRegister } from "./ZoweExplorerApiRegister";
@@ -43,7 +43,7 @@ const localize: nls.LocalizeFunc = nls.loadMessageBundle();
  * extensions to contribute their implementations.
  * @export
  */
-export class ZoweExplorerExtender implements ZoweExplorerApi.IApiExplorerExtender, ZoweExplorerTreeApi {
+export class ZoweExplorerExtender implements IApiExplorerExtender, ZoweExplorerTreeApi {
     public static ZoweExplorerExtenderInst: ZoweExplorerExtender;
 
     /**
@@ -146,7 +146,7 @@ export class ZoweExplorerExtender implements ZoweExplorerApi.IApiExplorerExtende
 
     /**
      *
-     * @implements ZoweExplorerApi.IApiExplorerExtender.initForZowe()
+     * @implements IApiExplorerExtender.initForZowe()
      * @param {string} profileType
      * @param {imperative.ICommandProfileTypeConfiguration[]} profileTypeConfigurations
      */
@@ -215,7 +215,7 @@ export class ZoweExplorerExtender implements ZoweExplorerApi.IApiExplorerExtende
     /**
      * Gives extenders access to the profiles loaded into memory by Zowe Explorer.
      *
-     * @implements ZoweExplorerApi.IApiExplorerExtender.getProfilesCache()
+     * @implements IApiExplorerExtender.getProfilesCache()
      * @returns {ProfilesCache}
      */
     public getProfilesCache(): ProfilesCache {
@@ -228,7 +228,7 @@ export class ZoweExplorerExtender implements ZoweExplorerApi.IApiExplorerExtende
      * to make them automatically appears in the Explorer drop-
      * down dialogs.
      *
-     * @implements ZoweExplorerApi.IApiExplorerExtender.reloadProfiles()
+     * @implements IApiExplorerExtender.reloadProfiles()
      * @param profileType optional profile type that the extender can specify
      */
     public async reloadProfiles(profileType?: string): Promise<void> {
